### Added

- d806fe4: Added phpstan to travis for static code analysis.

### Changed

- #1: PHP 7.2 min requirement and updated library to support it.
- #4: Added DBAL usage documentation and updated existing examples.
- 2d0951d: Applied PSR12 and additional rules to library.

### Removed

- #1: Removed support for PHP <7.2.
- d806fe4: Removed deprecated mongo db support.

### Fixed

- d806fe4: Fixed static code analysis issues.

+++ src/Spork/Batch/Strategy/ChunkStrategy.php
@@ -37,7 +37,7 @@ class ChunkStrategy extends AbstractStrategy
             $data = iterator_to_array($data);
         }
 
-        $size = ceil(count($data) / $this->forks);
+        $size = (int)ceil(count($data) / $this->forks);
 
         return array_chunk($data, $size, $this->preserveKeys);
     }
